[P4ADEV-3618] remove exportFiles API and tests; migrate to new mapping structure
